requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
black==22.1.0
click==8.0.4
Django==4.0.3
django-debug-toolbar==3.2.4
django-filter==21.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
gunicorn==20.1.0
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
whitenoise==6.1.0
